Updated SDL2 Submodule to Version 2.32.x Release Build
when UnleashedRecomp was released, it was running on SDL2 2.29.2 pre-release, this ultimately become 2.30.0. To closely align with the most up-to-date version of SDL2 and SDL2 Mixer: the submodule for both of them has been updated to be based on the 2.32.X branch.

Specifically: SDL Version 2.32.2-9-g2708e08f4, and SDL2_mixer 2.8.1.